refactor example's file structure
