+++ example/handlers/dummy.go
@@ -7,7 +7,7 @@ import (
 )
 
 // DummyRouter is router for test
-func DummyRouter() *mux.Router {
+func GetDummyRouter() *mux.Router {
 	router := mux.NewRouter()
 	router.HandleFunc("/v1/", func(w http.ResponseWriter, r *http.Request) {})
 	return router
