+++ example/handlers/main.go
@@ -1,0 +1,94 @@
+// This code is from https://www.codementor.io/codehakase/building-a-restful-api-with-golang-a6yivzqdo
+package handlers
+
+import (
+	"encoding/json"
+	"log"
+	"net/http"
+
+	"github.com/gorilla/mux"
+)
+
+// Person is the person type (more like an object)
+type Person struct {
+	ID        string   `json:"id,omitempty"`
+	Firstname string   `json:"firstname,omitempty"`
+	Lastname  string   `json:"lastname,omitempty"`
+	Address   *Address `json:"address,omitempty"`
+}
+
+// Address is the address type
+type Address struct {
+	City  string `json:"city,omitempty"`
+	State string `json:"state,omitempty"`
+}
+
+var people []Person
+
+// GetPeople displays all from the people var
+func GetPeople(w http.ResponseWriter, r *http.Request) {
+	json.NewEncoder(w).Encode(people)
+}
+
+// GetPerson displays a single data
+func GetPerson(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for _, item := range people {
+		if item.ID == params["id"] {
+			json.NewEncoder(w).Encode(item)
+			return
+		}
+	}
+	w.WriteHeader(http.StatusNotFound)
+}
+
+// CreatePerson creates a new item
+func CreatePerson(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	var person Person
+	_ = json.NewDecoder(r.Body).Decode(&person)
+	person.ID = params["id"]
+	people = append(people, person)
+	json.NewEncoder(w).Encode(person)
+}
+
+// CreatePersonFromForm creates a new item using form data
+func CreatePersonFromForm(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	_ = r.ParseForm()
+	person := &Person{}
+	person.ID = params["id"]
+	person.Firstname = r.Form.Get("firstname")
+	person.Lastname = r.Form.Get("lastname")
+
+	json.NewEncoder(w).Encode(person)
+}
+
+// DeletePerson deletes an item
+func DeletePerson(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for index, item := range people {
+		if item.ID == params["id"] {
+			people = append(people[:index], people[index+1:]...)
+			break
+		}
+	}
+}
+
+// main function to boot up everything
+func main() {
+
+	log.Fatal(http.ListenAndServe(":8000", GetRouter()))
+}
+
+func GetRouter() *mux.Router {
+	router := mux.NewRouter()
+	people = append(people, Person{ID: "1", Firstname: "John", Lastname: "Doe", Address: &Address{City: "City X", State: "State X"}})
+	people = append(people, Person{ID: "2", Firstname: "Koko", Lastname: "Doe", Address: &Address{City: "City Z", State: "State Y"}})
+	router.HandleFunc("/v1/people", GetPeople).Methods("GET")
+	router.HandleFunc("/v1/people/{id}", GetPerson).Methods("GET")
+	router.HandleFunc("/v1/people/{id}", CreatePerson).Methods("POST")
+	router.HandleFunc("/v1/form/people/{id}", CreatePersonFromForm).Methods("POST")
+	router.HandleFunc("/v1/people/{id}", DeletePerson).Methods("DELETE")
+	return router
+}
